requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
boto3==1.14.53
botocore==1.17.53
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-inmemorystorage==2.1.0
Django==3.1
django-allauth==0.42.0
django-environ==0.4.5
django-filter==2.3.0
django-rest-auth==0.9.5
django-storages==1.10
djangorestframework==3.11.1
docutils==0.15.2
drf-nested-routers==0.91
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
model-bakery==1.1.1
munch==2.5.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
